refactor - group components and screens for export
