+++ src/components/index.js
@@ -1,0 +1,6 @@
+export { default as AccountSelector } from './Account/AccountSelector';
+export { default as AccountButton } from './Account/AccountButton';
+export { default as FaucetButton  } from './Faucet/FaucetButton'
+export { default as InvitationDialog } from './Invitation/InvitationDialog'
+export { default as LiquidationAlert  } from './Liquidation/LiquidationAlert'
+export { default as TabBar } from './Tabbar/TabBar'
