+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { render } from 'react-dom';
 
 import App from './App';
 
-ReactDOM.render(<App />,
+render(<App />,
   document.getElementById('root')
 );
