+++ src/screens/Dashboard/DashboardPage.js
@@ -1,11 +1,11 @@
 import React from "react";
 
-import InvitationDialog from "components/Invitation/InvitationDialog";
-import LiquidationAlert from "components/Liquidation/LiquidationAlert";
+import { InvitationDialog, LiquidationAlert } from "components";
 import { useSubstrate } from "services/substrate-lib";
 import Dashboard from "./Dashboard";
 import Wallet from "./Wallet";
 import { WalletContextProvider } from "./WalletContext";
+
 import "./DashboardPage.css";
 
 export default function Main(props) {
