+++ src/screens/index.js
@@ -1,0 +1,5 @@
+export { default as ConnectPage } from './Connect/ConnectPage';
+export { default as DashboardPage } from './Dashboard/DashboardPage';
+export { default as ExchangePage  } from './Exchange/ExchangePage'
+export { default as MarketLists } from './Invest/MarketLists'
+export { default as WelcomePage  } from './Welcome/WelcomePage'
